backend/pyproject.toml
[project]
name = "ai-mock-interview-platform-pending"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=1.1.1",
    "google-genai>=1.19.0",
    "kokoro>=0.9.4",
    "librosa>=0.11.0",
    "numpy==2.2",
    "openai-whisper>=20240930",
    "pip>=25.1.1",
    "pyaudio>=0.2.14",
    "pyttsx3>=2.98",
    "silero-vad>=5.1.2",
    "sounddevice>=0.5.2",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
]
